requirements.txt
scikit-learn
pandas
numpy
django
ipykernel
pyYAML
matplotlib
seaborn
-e .